worktrees/main/pyproject.toml
[project]
name = "yazot-mcp-server"
version = "0.1.0"
description = "Yet Another Zotero MCP server"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=0.1.0",
    "pyzotero>=1.5.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "pypdf2>=3.0.0",
    "aiofiles>=23.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
    "black>=24.8.0",
    "pre-commit>=3.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM", "RUF"]
ignore = ["E501"]
target-version = "py312"

[tool.ruff.per-file-ignores]
"tests/*" = ["S101"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
disallow_untyped_defs = true
disallow_any_unimported = false
no_implicit_optional = true
check_untyped_defs = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "normal"

[[tool.mypy.overrides]]
module = "pydantic"
ignore_missing_imports = false

[[tool.mypy.overrides]]
module = "pydantic.*"
ignore_missing_imports = false

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.black]
line-length = 100
target-version = ['py312']

[tool.uv]
dev-dependencies = [
    "ipython>=8.0.0",
]
